requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pritunl-api
outline-vpn-api
